Now using e.Matches where appropriate. Renamed most things from k8s to kubernetes.
