+++ main.go
@@ -7,7 +7,7 @@ import (
 	_ "github.com/codeamp/circuit/plugins/githubstatus"
 	_ "github.com/codeamp/circuit/plugins/gitsync"
 	_ "github.com/codeamp/circuit/plugins/heartbeat"
-	_ "github.com/codeamp/circuit/plugins/k8s"
+	_ "github.com/codeamp/circuit/plugins/kubernetes"
 	_ "github.com/codeamp/circuit/plugins/route53"
 )
