+++ plugins/kubernetes/k8s_test.go
@@ -1,15 +1,17 @@
-package k8s_test
+package kubernetes_test
 
 import (
 	"bytes"
 	"errors"
+	"fmt"
+	"os"
+	"path"
 	"strings"
 	"testing"
 	"time"
 
 	"github.com/codeamp/circuit/plugins"
-	"github.com/codeamp/circuit/plugins/k8s"
-	"github.com/codeamp/circuit/plugins/k8s/testdata"
+	"github.com/codeamp/circuit/plugins/kubernetes"
 	log "github.com/codeamp/logger"
 	"github.com/codeamp/transistor"
 	"github.com/spf13/viper"
@@ -24,7 +26,7 @@ type TestSuite struct {
 
 var viperConfig = []byte(`
 plugins:
-  k8s:
+  kubernetes:
     workers: 1
 `)
 
@@ -35,13 +37,13 @@ func (suite *TestSuite) SetupSuite() {
 	viper.AutomaticEnv()
 	viper.ReadConfig(bytes.NewBuffer(viperConfig))
 
-	transistor.RegisterPlugin("k8s", func() transistor.Plugin {
-		return &k8s.K8s{}
+	transistor.RegisterPlugin("kubernetes", func() transistor.Plugin {
+		return &kubernetes.Kubernetes{}
 	})
 
 	config := transistor.Config{
 		Plugins:        viper.GetStringMap("plugins"),
-		EnabledPlugins: []string{"k8s"},
+		EnabledPlugins: []string{"kubernetes"},
 	}
 
 	ag, _ := transistor.NewTestTransistor(config)
@@ -50,7 +52,7 @@ func (suite *TestSuite) SetupSuite() {
 }
 
 // func (suite *TestSuite) TestCleanupLBOffice() {
-// 	suite.transistor.Events <- testdata.LBTCPEvent(plugins.GetAction("destroy"), plugins.GetType("office"))
+// 	suite.transistor.Events <- LBTCPEvent(plugins.GetAction("destroy"), plugins.GetType("office"))
 
 // 	e := suite.transistor.GetTestEvent("plugins.ProjectExtension:status", 60)
 // 	assert.Equal(suite.T(), plugins.GetState("deleted"), e.Payload.(plugins.ProjectExtension).State, e.Payload.(plugins.ProjectExtension).StateMessage)
@@ -66,7 +68,7 @@ func (suite *TestSuite) TestLBTCPOffice() {
 		log.Fatal("TestLBTCPOffice: Test timeout")
 	}()
 
-	suite.transistor.Events <- testdata.LBTCPEvent(plugins.GetAction("update"), plugins.GetType("office"))
+	suite.transistor.Events <- LBTCPEvent(plugins.GetAction("update"), plugins.GetType("office"))
 
 	e := suite.transistor.GetTestEvent("plugins.ProjectExtension:status", 120)
 	assert.Equal(suite.T(), plugins.GetState("complete"), e.Payload.(plugins.ProjectExtension).State, e.Payload.(plugins.ProjectExtension).StateMessage)
@@ -74,7 +76,7 @@ func (suite *TestSuite) TestLBTCPOffice() {
 		return
 	}
 
-	testdata.LBTCPEvent(plugins.GetAction("update"), plugins.GetType("office"))
+	LBTCPEvent(plugins.GetAction("update"), plugins.GetType("office"))
 
 	for {
 		e = suite.transistor.GetTestEvent("plugins.ProjectExtension:status", 120)
@@ -83,7 +85,7 @@ func (suite *TestSuite) TestLBTCPOffice() {
 		}
 	}
 
-	suite.transistor.Events <- testdata.LBTCPEvent(plugins.GetAction("destroy"), plugins.GetType("office"))
+	suite.transistor.Events <- LBTCPEvent(plugins.GetAction("destroy"), plugins.GetType("office"))
 
 	e = suite.transistor.GetTestEvent("plugins.ProjectExtension:status", 10)
 	assert.Equal(suite.T(), string(plugins.GetState("deleted")), string(e.Payload.(plugins.ProjectExtension).State))
@@ -107,9 +109,9 @@ func (suite *TestSuite) TestBasicSuccessDeploy() {
 	defer timer.Stop()
 
 	var e transistor.Event
-	suite.transistor.Events <- testdata.BasicReleaseEvent()
+	suite.transistor.Events <- BasicReleaseEvent()
 	e = suite.transistor.GetTestEvent("plugins.ReleaseExtension:status", 5)
-	assert.Equal(suite.T(), string(plugins.GetState("running")), string(e.Payload.(plugins.ReleaseExtension).State))
+	assert.Equal(suite.T(), plugins.GetState("running"), e.Payload.(plugins.ReleaseExtension).State)
 
 	go func() {
 		<-timer.C
@@ -124,7 +126,7 @@ func (suite *TestSuite) TestBasicSuccessDeploy() {
 	}
 
 	suite.T().Log(e.Payload.(plugins.ReleaseExtension).StateMessage)
-	assert.Equal(suite.T(), string(plugins.GetState("complete")), string(e.Payload.(plugins.ReleaseExtension).State))
+	assert.Equal(suite.T(), plugins.GetState("complete"), e.Payload.(plugins.ReleaseExtension).State)
 }
 
 func (suite *TestSuite) TestBasicFailedDeploy() {
@@ -137,10 +139,10 @@ func (suite *TestSuite) TestBasicFailedDeploy() {
 	}()
 
 	var e transistor.Event
-	suite.transistor.Events <- testdata.BasicFailedReleaseEvent()
+	suite.transistor.Events <- BasicFailedReleaseEvent()
 
 	e = suite.transistor.GetTestEvent("plugins.ReleaseExtension:status", 5)
-	assert.Equal(suite.T(), string(plugins.GetState("running")), string(e.Payload.(plugins.ReleaseExtension).State))
+	assert.Equal(suite.T(), plugins.GetState("running"), e.Payload.(plugins.ReleaseExtension).State)
 
 	for {
 		e = suite.transistor.GetTestEvent("plugins.ReleaseExtension:status", 30)
@@ -150,7 +152,7 @@ func (suite *TestSuite) TestBasicFailedDeploy() {
 	}
 
 	suite.T().Log(e.Payload.(plugins.ReleaseExtension).StateMessage)
-	assert.Equal(suite.T(), string(plugins.GetState("failed")), string(e.Payload.(plugins.ReleaseExtension).State))
+	assert.Equal(suite.T(), plugins.GetState("failed"), e.Payload.(plugins.ReleaseExtension).State)
 }
 
 func TestDeployments(t *testing.T) {
@@ -179,7 +181,7 @@ func (suite *TestSuite) TearDownSuite() {
 }
 
 func verifyDeploymentArtifacts() error {
-	e := testdata.BasicReleaseEvent()
+	e := BasicReleaseEvent()
 
 	basicReleaseEventArtifacts := map[string]string{
 		"user":                  "",
@@ -205,7 +207,7 @@ func verifyDeploymentArtifacts() error {
 }
 
 func verifyLoadBalancerArtifacts() error {
-	e := testdata.LBTCPEvent(plugins.GetAction("update"), plugins.GetType("office"))
+	e := LBTCPEvent(plugins.GetAction("update"), plugins.GetType("office"))
 
 	lbTCPArtifacts := map[string]string{
 		"service":               "",
@@ -229,4 +231,189 @@ func verifyLoadBalancerArtifacts() error {
 	}
 
 	return nil
+}
+
+func GetCreateProjectExtension() plugins.ProjectExtension {
+	d := GetBasicProjectExtension()
+	d.Action = plugins.GetAction("create")
+	d.State = plugins.GetState("waiting")
+	return d
+}
+
+func GetDestroyProjectExtension() plugins.ProjectExtension {
+	d := GetBasicProjectExtension()
+	d.Action = plugins.GetAction("destroy")
+	d.State = plugins.GetState("waiting")
+	return d
+}
+
+func GetBasicProjectExtension() plugins.ProjectExtension {
+	var kubeconfig string
+	if kubeconfig = os.Getenv("KUBECONFIG"); kubeconfig == "" {
+		kubeconfig = path.Join(os.Getenv("HOME"), ".kube", "config")
+	}
+
+	extensionEvent := plugins.ProjectExtension{
+		Slug:        "kubernetesloadbalancers",
+		Environment: "testing",
+		Action:      plugins.GetAction("create"),
+		State:       plugins.GetState("waiting"),
+		Project: plugins.Project{
+			Repository: "checkr/deploy-test",
+			// Services: []plugins.Service{
+			// 	plugins.Service{},
+			// },
+		},
+	}
+
+	return extensionEvent
+}
+
+func LBDataForTCP(action plugins.Action, t plugins.Type) plugins.ProjectExtension {
+	project := plugins.Project{
+		Repository: "checkr/nginx-test-success",
+	}
+
+	lbe := plugins.ProjectExtension{
+		Slug:        "kubernetesloadbalancers",
+		Action:      action,
+		Environment: "testing",
+		Project:     project,
+		ID:          "nginx-test-lb-asdf1234",
+	}
+	return lbe
+}
+
+func LBTCPEvent(action plugins.Action, t plugins.Type) transistor.Event {
+	data := LBDataForTCP(action, t)
+	event := transistor.NewEvent(data, nil)
+
+	var kubeConfigPath string
+	if kubeConfigPath = os.Getenv("KUBECONFIG_PATH"); kubeConfigPath == "" {
+		kubeConfigPath = path.Join(os.Getenv("HOME"), ".kube", "config")
+	}
+
+	event.AddArtifact("service", "nginx-test-service-asdf", false)
+	event.AddArtifact("name", "nginx-test-lb-asdf1234", false)
+	event.AddArtifact("ssl_cert_arn", "arn:1234:arnid", false)
+	event.AddArtifact("access_log_s3_bucket", "test-s3-logs-bucket", false)
+	event.AddArtifact("type", fmt.Sprintf("%v", t), false)
+
+	// For Kube connectivity
+	event.AddArtifact("kubeconfig", kubeConfigPath, false)
+	event.AddArtifact("client_certificate", "", false)
+	event.AddArtifact("client_key", "", false)
+	event.AddArtifact("certificate_authority", "", false)
+
+	var listener_pairs []interface{} = make([]interface{}, 2, 2)
+	listener_pairs[0] = map[string]interface{}{
+		"serviceProtocol": "TCP",
+		"port":            "443",
+		"containerPort":   "3000",
+	}
+	listener_pairs[1] = map[string]interface{}{
+		"serviceProtocol": "TCP",
+		"port":            "444",
+		"containerPort":   "3001",
+	}
+	event.AddArtifact("listener_pairs", listener_pairs, false)
+
+	return event
+}
+
+func BasicFailedReleaseEvent() transistor.Event {
+	extension := BasicReleaseExtension()
+	extension.Release.Services[0].Command = "/bin/false"
+
+	event := transistor.NewEvent(extension, nil)
+	addBasicReleaseExtensionArtifacts(extension, &event)
+
+	return event
+}
+
+func addBasicReleaseExtensionArtifacts(extension plugins.ReleaseExtension, event *transistor.Event) {
+	var kubeConfigPath string
+	if kubeConfigPath = os.Getenv("KUBECONFIG_PATH"); kubeConfigPath == "" {
+		kubeConfigPath = path.Join(os.Getenv("HOME"), ".kube", "config")
+	}
+
+	event.AddArtifact("user", "test", false)
+	event.AddArtifact("password", "test", false)
+	event.AddArtifact("host", "test", false)
+	event.AddArtifact("email", "test", false)
+	event.AddArtifact("image", "nginx", false)
+
+	for idx := range event.Artifacts {
+		event.Artifacts[idx].Source = "dockerbuilder"
+	}
+
+	event.AddArtifact("kubeconfig", kubeConfigPath, false)
+	event.AddArtifact("client_certificate", "", false)
+	event.AddArtifact("client_key", "", false)
+	event.AddArtifact("certificate_authority", "", false)
+}
+
+func BasicReleaseEvent() transistor.Event {
+	extension := BasicReleaseExtension()
+
+	event := transistor.NewEvent(extension, nil)
+	addBasicReleaseExtensionArtifacts(extension, &event)
+
+	return event
+}
+
+func BasicReleaseExtension() plugins.ReleaseExtension {
+
+	deploytestHash := "4930db36d9ef6ef4e6a986b6db2e40ec477c7bc9"
+
+	release := plugins.Release{
+		Project: plugins.Project{
+			Repository: "checkr/deploy-test",
+		},
+		Git: plugins.Git{
+			Url:           "https://github.com/checkr/deploy-test.git",
+			Protocol:      "HTTPS",
+			Branch:        "master",
+			RsaPrivateKey: "",
+			RsaPublicKey:  "",
+			Workdir:       "/tmp/something",
+		},
+		Services: []plugins.Service{
+			plugins.Service{
+				Name: "www",
+				Listeners: []plugins.Listener{
+					{
+						Port:     80,
+						Protocol: "TCP",
+					},
+				},
+				State: plugins.GetState("waiting"),
+				Spec: plugins.ServiceSpec{
+
+					CpuRequest:                    "10m",
+					CpuLimit:                      "500m",
+					MemoryRequest:                 "1Mi",
+					MemoryLimit:                   "500Mi",
+					TerminationGracePeriodSeconds: int64(1),
+				},
+				Replicas: 1,
+			},
+		},
+		HeadFeature: plugins.Feature{
+			Hash:       deploytestHash,
+			ParentHash: deploytestHash,
+			User:       "",
+			Message:    "Test",
+		},
+		Environment: "testing",
+	}
+
+	releaseExtension := plugins.ReleaseExtension{
+		Slug:    "kubernetesdeployments",
+		Action:  plugins.GetAction("create"),
+		State:   plugins.GetState("waiting"),
+		Release: release,
+	}
+
+	return releaseExtension
 }
